# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  controllers.SwaggerLogPenyakitTanaman:
    description: Log penyakit terkait tanaman
    properties:
      catatan:
        type: string
      created_at:
        type: string
      foto:
        type: string
      foto_log_penyakit_id:
        type: string
      id:
        type: integer
      kondisi:
        type: string
      penyakit_id:
        type: integer
      saran_perawatan:
        type: string
      tanaman_id:
        type: integer
      updated_at:
        type: string
    type: object
...
  controllers.UpdateKebunRequest:
    properties:
      mdpl:
        type: string
      nama_kebun:
        type: string
    type: object
...
  utils.Pagination:
    properties:
      has_next:
        type: boolean
      has_prev:
        type: boolean
      page:
        type: integer
      per_page:
        type: integer
      total_pages:
        type: integer
      total_rows:
        type: integer
    type: object
  utils.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      meta: {}
      success:
        type: boolean
    type: object
host: localhost:2005
info:
  contact: {}
  description: Dokumentasi API Avocycle (CRUD, Pagination, dan Response Format)
  title: Avocycle API
  version: "1.0"
paths:
  /Log-Penyakit-Tanaman:
    get:
      consumes:
      - application/json
      description: Mendapatkan daftar semua log penyakit tanaman dengan pagination
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Berhasil mengambil data
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/controllers.SwaggerLogPenyakitTanaman'
                  type: array
                meta:
                  $ref: '#/definitions/utils.Pagination'
              type: object
        "500":
          description: Gagal koneksi atau query database
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Get all log penyakit tanaman
      tags:
      - LogPenyakitTanaman
  /Log-Penyakit-Tanaman/{id}:
    get:
      consumes:
      - application/json
      description: Mendapatkan detail log penyakit tanaman berdasarkan ID
      parameters:
      - description: ID Log Penyakit Tanaman
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Detail ditemukan
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/controllers.SwaggerLogPenyakitTanaman'
              type: object
        "404":
          description: Data tidak ditemukan
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Gagal retrieve data
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Get log penyakit tanaman by ID
      tags:
      - LogPenyakitTanaman
  /Log-Penyakit-Tanaman/Tanaman/{id_tanaman}:
    get:
      consumes:
      - application/json
      description: Mendapatkan semua log penyakit berdasarkan tanaman_id dengan pagination
      parameters:
      - description: ID Tanaman
        in: path
        name: id_tanaman
        required: true
        type: integer
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Berhasil mengambil data
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/controllers.SwaggerLogPenyakitTanaman'
                  type: array
                meta:
                  $ref: '#/definitions/utils.Pagination'
              type: object
        "404":
          description: Data tidak ditemukan
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Gagal koneksi atau query database
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Get log penyakit tanaman by Tanaman ID
      tags:
      - LogPenyakitTanaman
...
  /kebun:
    get:
      description: Mengambil daftar kebun dengan pagination (menggunakan meta pagination
        sesuai utils Pagination)
      parameters:
      - description: Halaman
        in: query
        name: page
        type: integer
      - description: Jumlah data per halaman
        in: query
        name: per_page
        type: integer
      responses:
        "200":
          description: Data kebun berhasil diambil
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Ambil semua kebun
      tags:
      - Kebun
    post:
      consumes:
      - application/json
      description: Membuat data kebun baru (response mengikuti utils.Response)
      parameters:
      - description: Input kebun
        in: body
        name: request
        required: true
        schema:
          properties:
            mdpl:
              type: string
            nama_kebun:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - Bearer: []
      summary: Membuat kebun baru
      tags:
      - Kebun
  /kebun/{id}:
    delete:
      description: Menghapus data kebun berdasarkan ID
      parameters:
      - description: ID Kebun
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - Bearer: []
      summary: Hapus kebun
      tags:
      - Kebun
    get:
      description: Mengambil detail kebun berdasarkan ID
      parameters:
      - description: ID Kebun
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Ambil detail kebun
      tags:
      - Kebun
    put:
      consumes:
      - application/json
      description: Mengupdate data kebun berdasarkan ID (partial update)
      parameters:
      - description: ID Kebun
        in: path
        name: id
        required: true
        type: integer
      - description: Input update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.UpdateKebunRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - Bearer: []
      summary: Update kebun
      tags:
      - Kebun
...
schemes:
- http
securityDefinitions:
  Bearer:
    description: Type "Bearer <your-token>"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"